feat: migrate AppFunctions to self-hosted alpha10 AppFunctionService
androidx.appfunctions alpha10 (July 2026) stopped publishing the
appfunctions-service artifact this app depended on for its "add word"
AppFunction. The Add AppFunctions API guide now expects apps to host
their own service instead:

- VocabularyFunctions (a plain Hilt-injected class registered via
  AppFunctionConfiguration.Provider in MyApp) is replaced by
  BaseVocabularyAppFunctionService, an abstract class extending
  AppFunctionService, annotated @AndroidEntryPoint +
  @AppFunctionServiceEntryPoint(serviceName = "VocabularyAppFunctionService").
  Dependencies move from constructor injection to Hilt field injection
  (services can't take constructor params), and the AppFunctionContext
  parameter is dropped from addWord() since alpha10 no longer passes one
  (the service itself is a Context).
- MyApp drops AppFunctionConfiguration.Provider entirely - service
  registration now happens through the manifest instead.
- AndroidManifest.xml declares the KSP-generated concrete
  VocabularyAppFunctionService with BIND_APP_FUNCTION_SERVICE permission
  and the schema/v2 metadata properties KSP generates
  (app_functions_schema.xsd, vocabulary_app_function_service.xml).
- Drops the appfunctions-service dependency and the now-obsolete
  appfunctions:aggregateAppFunctions KSP arg (function discovery is
  scoped to the annotated service class now, not aggregated project-wide).

Verified past compilation: ./gradlew assembleDebug confirms the KSP
compiler generates VocabularyAppFunctionService in the expected package,
and unzipping the built debug APK confirms both
assets/app_functions_schema.xsd and assets/vocabulary_app_function_service.xml
are packaged in, matching the manifest's <property> declarations.
